broke out more functions
