+++ src/components/N64Configuration.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useEffect } from "react";
 import ChromeSamples from "../utils/ChromeSamples";
-import { brUuid, mtu, block, pakSize } from "../utils/constants";
+import { brUuid, pakSize } from "../utils/constants";
 import Button from "@mui/material/Button";
 import ProgressBar from "react-bootstrap/ProgressBar";
 import { useFilePicker } from "use-file-picker";
@@ -22,12 +22,15 @@ import Paper from "@mui/material/Paper";
 import Stack from "@mui/material/Stack";
 import { Divider, Typography } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-
-var cancel = 0;
+import downloadFile from "../utils/downloadFile";
+import saveGlobalCfg from "../utils/saveGlobalCfg";
+import makeFormattedPak from "../utils/makeFormattedPak";
+import n64WriteFile from "../utils/n64WriteFile";
 
 function N64Configuration(props) {
   const navigate = useNavigate();
   const startTime = useRef(0);
+  const cancel = useRef(0);
   const [show, setShow] = useState(false);
   const handleClose = () => setShow(false);
   const handleFormat = () => setShow(true);
@@ -86,37 +89,32 @@ function N64Configuration(props) {
         setShowCancel(false);
         setShowProgress(false);
         setShowButtons(true);
-        cancel = 0;
+        cancel.current = 0;
       });
   };
 
   const pakWrite = (evt) => {
-    writeFile(filesContent[0].content.slice(0, pakSize));
+    setShowButtons(false);
+    setShowProgress(true);
+    setShowCancel(true);
+    startTime.current = performance.now();
+    n64WriteFile(props.btService, filesContent[0].content.slice(0, pakSize), pak, setProgress, cancel).then((_) => {
+      clear();
+      setShowButtons(true);
+      setShowCancel(false);
+      setShowProgress(false);
+    })
+    .catch((error) => {
+      ChromeSamples.log("Argh! " + error);
+      setShowButtons(true);
+      setShowCancel(false);
+      setShowProgress(false);
+      cancel.current = 0;
+    });
   };
 
-  // Source: https://newbedev.com/saving-binary-data-as-file-using-javascript-from-a-browser
-  function downloadFile(blob, filename) {
-    var url = window.URL.createObjectURL(blob);
-
-    var a = document.createElement("a");
-    a.style = "display: none";
-    a.href = url;
-    a.download = filename;
-
-    document.body.appendChild(a);
-    a.click();
-
-    document.body.removeChild(a);
-
-    // On Edge, revokeObjectURL should be called only after
-    // a.click() has completed, atleast on EdgeHTML 15.15048
-    setTimeout(function () {
-      window.URL.revokeObjectURL(url);
-    }, 1000);
-  }
-
   const abortFileTransfer = () => {
-    cancel = 1;
+    cancel.current = 1;
   };
 
   const transferProgress = (total, loaded) => {
@@ -132,62 +130,27 @@ function N64Configuration(props) {
   const pakFormat = (evt) => {
     handleClose();
     setProgress(0);
-    function writeAt(ofs) {
-      for (let i = 0; i < 32; i++) data[ofs + i] = block[i];
-    }
-
-    const data = new Uint8Array(32768);
-    const block = new Uint8Array(32);
-
-    // generate id block
-    block[1] = 0 | ((Math.random() * 256) & 0x3f);
-    block[5] = 0 | ((Math.random() * 256) & 0x7);
-    block[6] = 0 | (Math.random() * 256);
-    block[7] = 0 | (Math.random() * 256);
-    block[8] = 0 | ((Math.random() * 256) & 0xf);
-    block[9] = 0 | (Math.random() * 256);
-    block[10] = 0 | (Math.random() * 256);
-    block[11] = 0 | (Math.random() * 256);
-    block[25] = 0x01; // device bit
-    block[26] = 0x01; // bank size int (must be exactly '01')
-
-    // calculate pakId checksum
-    let sumA = 0,
-      sumB = 0xfff2;
-    for (let i = 0; i < 28; i += 2) {
-      sumA += (block[i] << 8) + block[i + 1];
-      sumA &= 0xffff;
-    }
-    sumB -= sumA;
-    // store checksums
-    block[28] = sumA >> 8;
-    block[29] = sumA & 0xff;
-    block[30] = sumB >> 8;
-    block[31] = sumB & 0xff;
-
-    // write checksum block to multiple sections in header page
-    writeAt(32);
-    writeAt(96);
-    writeAt(128);
-    writeAt(192);
-
-    // init IndexTable and backup (plus checksums)
-    for (let i = 5; i < 128; i++) {
-      data[256 + i * 2 + 1] = 3;
-      data[512 + i * 2 + 1] = 3;
-    }
-    data[257] = 0x71;
-    data[513] = 0x71;
-
-    //for(let i = 0; i < 32; i++) data[i] = i; // write label - needs to be verified
-    //data[0] = 0x81; // libultra's 81 mark
-
-    writeFile(data.buffer);
+    setShowButtons(false);
+    setShowProgress(true);
+    setShowCancel(true);
+    startTime.current = performance.now();
+    n64WriteFile(props.btService, makeFormattedPak().buffer, pak, setProgress, cancel).then((_) => {
+      setShowButtons(true);
+      setShowCancel(false);
+      setShowProgress(false);
+    })
+    .catch((error) => {
+      ChromeSamples.log("Argh! " + error);
+      setShowButtons(true);
+      setShowCancel(false);
+      setShowProgress(false);
+      cancel.current = 0;
+    });
   };
 
   const readRecursive = (chrc, data, offset) => {
     return new Promise(function (resolve, reject) {
-      if (cancel === 1) {
+      if (cancel.current === 1) {
         throw new Error("Cancelled");
       }
       transferProgress(pakSize, offset);
@@ -208,40 +171,6 @@ function N64Configuration(props) {
                 " sec"
             );
             resolve(data);
-          }
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    });
-  };
-
-  const writeRecursive = (chrc, data, offset) => {
-    return new Promise(function (resolve, reject) {
-      var curBlock = ~~(offset / block) + 1;
-      if (cancel === 1) {
-        throw new Error("Cancelled");
-      }
-      transferProgress(data.byteLength, offset);
-      let tmpViewSize = curBlock * block - offset;
-      if (tmpViewSize > mtu) {
-        tmpViewSize = mtu;
-      }
-      var tmpView = new DataView(data, offset, tmpViewSize);
-      chrc
-        .writeValue(tmpView)
-        .then((_) => {
-          offset += tmpViewSize;
-          if (offset < data.byteLength) {
-            resolve(writeRecursive(chrc, data, offset));
-          } else {
-            setProgress(100);
-            ChromeSamples.log(
-              "File upload done. Took: " +
-                Math.round(performance.now() - startTime.current) / 1000 +
-                " sec"
-            );
-            resolve();
           }
         })
         .catch((error) => {
@@ -282,70 +211,6 @@ function N64Configuration(props) {
         });
     });
   };
-
-  const writeFile = (data) => {
-    setShowButtons(false);
-    setShowProgress(true);
-    setShowCancel(true);
-    var offset = new Uint32Array(1);
-    let ctrl_chrc = null;
-
-    props.btService
-      .getCharacteristic(brUuid[10])
-      .then((chrc) => {
-        ctrl_chrc = chrc;
-        offset[0] = Number(pak) * pakSize;
-        return ctrl_chrc.writeValue(offset);
-      })
-      .then((_) => {
-        return props.btService.getCharacteristic(brUuid[11]);
-      })
-      .then((chrc) => {
-        startTime.current = performance.now();
-        setShowButtons(false);
-        return writeRecursive(chrc, data, 0);
-      })
-      .then((_) => {
-        offset[0] = 0;
-        return ctrl_chrc.writeValue(offset);
-      })
-      .then((_) => {
-        clear();
-        setShowButtons(true);
-        setShowCancel(false);
-        setShowProgress(false);
-      })
-      .catch((error) => {
-        ChromeSamples.log("Argh! " + error);
-        setShowButtons(true);
-        setShowCancel(false);
-        setShowProgress(false);
-        cancel = 0;
-      });
-  };
-
-  function saveGlobal(memoryBank) {
-    var data = props.globalCfg;
-    data[3] = memoryBank - 1;
-    return new Promise(function (resolve, reject) {
-      ChromeSamples.log("Get Global Config CHRC...");
-      props.btService
-        .getCharacteristic(brUuid[1])
-        .then((chrc) => {
-          ChromeSamples.log("Writing Global Config...");
-          return chrc.writeValue(data);
-        })
-        .then((_) => {
-          ChromeSamples.log("Global Config saved");
-          resolve();
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    });
-  }
-
-  
 
   return (
     <Paper
@@ -492,7 +357,9 @@ function N64Configuration(props) {
                     id="btnPakRead"
                     variant="outlined"
                     onClick={() => {
-                      saveGlobal(bank);
+                      let globalCfg = props.globalCfg;
+                      globalCfg[3] = bank - 1;
+                      saveGlobalCfg(props.btService, globalCfg);
                     }}
                   >
                     Save
@@ -500,8 +367,6 @@ function N64Configuration(props) {
                 </Stack>
               </AccordionDetails>
             </Accordion>
-
-
           </Stack>
         )}
         <Box>
