+++ src/utils/downloadFile.js
@@ -1,0 +1,21 @@
+export function downloadFile(blob, filename) {
+    var url = window.URL.createObjectURL(blob);
+
+    var a = document.createElement("a");
+    a.style = "display: none";
+    a.href = url;
+    a.download = filename;
+
+    document.body.appendChild(a);
+    a.click();
+
+    document.body.removeChild(a);
+
+    // On Edge, revokeObjectURL should be called only after
+    // a.click() has completed, atleast on EdgeHTML 15.15048
+    setTimeout(function () {
+      window.URL.revokeObjectURL(url);
+    }, 1000);
+  }
+
+  export default downloadFile;
